Change send_backspaces/string to change_string
change_string can easily be implemented by the same combination of
send_backspaces and send_string that was used before. The advantage of
outputting change_string(before, after) is that it gives the output
class the chance to reject the change if the 'before' string does not
match the string adjacent to the cursor.

For example, if the cursor is moved outside of Plover's control, then
currently pressing * will blindly delete the wrong text.

(this change is intended to support use with the IBus input method)