Allow listeners to reject translations
If a Translator's listener raises TranslationNotPossible, the translator
will reset its state and try again. This will not happen in the current
implementation of Plover, when the output is done by sending keypress
events. Other output methods may have more information about where the
cursor is, so may want to reject translations if, for example, the
cursor has moved outside of Plover's control.

(this change is intended to support use with the IBus input method)